refactor(builder): delete dead active-row CSS and relocate RibDecorations
Removes the entire @container slot / [data-layout='2x3-vertical'] /
[data-calc-open] CSS block that drove the legacy active-row layout
switching. With CompactRow + GridRow handling everything via Tailwind
viewport breakpoints, none of these rules are reachable from production:

- .row-active, .row-vertical-content, .row-right (display:contents
  wrappers used by ResizeObserver-driven layout flips)
- .rib, .rib-right, .rib-remove-fallback (RIB orientation/visibility)
- .rib-decorations (always-hidden when collapsed, flex when vertical)

styles/globals.css: 329 -> 235 lines. The unrelated calc-col-enter
keyframe is preserved.

Also relocates lanes/rib-decorations.tsx to _deprecated/ - it had no
production callers (only its own test rendered it). Inert
className='rib-decorations' string left in place; the CSS rule it
depended on is deleted but the component itself is dead.

Two attribute writers (data-layout, data-calc-open) remain in
team-workspace.tsx as no-ops; Phase 1g rips them along with the
[[data-density=compact]_&]: Tailwind variants.

Tests: 90 suites, 1963 pass. Typecheck clean.
