+++ apps/web/src/styles/globals.css
@@ -221,100 +221,6 @@
   box-shadow: inset 0 0 0 2px currentcolor;
 }
 
-/* -- Active row @container responsive layout ------------------------------ */
-
-.row-right {
-  display: contents;
-}
-
-/* When calc is off the slot max-width shrinks below 1579px, which would
-   normally trigger the mid-stack layout. Override to keep side-by-side. */
-[data-calc-open="false"] .row-right {
-  display: contents;
-}
-
-@container slot (max-width: 1579px) {
-  .row-right {
-    display: flex;
-    flex-direction: column;
-    flex: 1 1 0;
-    min-width: 0;
-  }
-}
-
-.row-vertical-content {
-  display: contents;
-}
-
-[data-layout="2x3-vertical"] .row-vertical-content {
-  display: flex;
-  flex-direction: column;
-  flex: 1 1 0;
-  min-width: 0;
-}
-
-[data-layout="2x3-vertical"] .row-active {
-  flex-direction: column;
-  align-items: stretch;
-}
-
-[data-layout="2x3-vertical"] .rib {
-  flex-direction: row;
-  justify-content: space-between;
-  width: 100%;
-  padding: 4px 10px;
-  border-right: none;
-  border-bottom: 1px dashed var(--border);
-}
-
-@container slot (max-width: 699px) {
-  .row-active {
-    flex-direction: column;
-    align-items: stretch;
-  }
-
-  .rib {
-    flex-direction: row;
-    justify-content: space-between;
-    width: 100%;
-    padding: 4px 10px;
-    border-right: none;
-    border-bottom: 1px dashed var(--border);
-  }
-
-  .rib-right {
-    display: none;
-  }
-
-  .rib-remove-fallback {
-    display: flex;
-  }
-
-  .rib-decorations {
-    display: none;
-  }
-}
-
-[data-layout="2x3-vertical"] .rib-right {
-  display: none;
-}
-
-[data-layout="2x3-vertical"] .rib-remove-fallback {
-  display: flex;
-}
-
-[data-layout="2x3-vertical"] .rib-decorations {
-  display: flex;
-  flex-direction: row;
-  gap: 4px;
-}
-
-.rib-remove-fallback {
-  display: none;
-}
-
-
-
 /* -- Calc column enter animation (keyframe for Tailwind animate-[]) ------- */
 
 @keyframes calc-col-enter {
